+++ lib/integrate.dart
@@ -1,0 +1,11 @@
+/// Numerical methods for integrating extension
+///
+/// Following methods available:
+/// - By rectangles (left, middle, right)
+/// - Simpson's method
+/// - By trapezoids
+///
+/// In order to use the functionality import the `integrate.dart` file
+library math_parser_integrate;
+
+export 'src/extensions/integrate.dart';