+++ lib/src/parse_errors.dart
@@ -1,8 +1,13 @@
-/// Math parsing error
+import 'package:math_parser/src/math_errors.dart';
+
+/// Math Parse Exception
 ///
-/// Errors must extends this class
-abstract class MathException implements Exception {
-  const MathException();
+/// All sorts of errors related to parsing a string to [MathNode] used by
+/// [MathNodeExpression]. All actuall exceptions are being extended from this
+/// class.
+abstract class MathParseException extends MathException {
+  /// Creates a new Math Parse Exception
+  const MathParseException();
 }
 
 /// Missing Operator Operand Exception
@@ -10,10 +15,11 @@ abstract class MathException implements Exception {
 /// Thrown when an operator lacks its left or right operand.
 ///
 /// The operator which caused the problem is stored in [operator]
-class MissingOperatorOperandException extends MathException {
+class MissingOperatorOperandException extends MathParseException {
   @override
   String toString() {
-    return 'MissingOperatorOperandException: "$operator" has insufficient neighboring expressions';
+    return 'MissingOperatorOperandException: "$operator" has insufficient '
+        'neighboring expressions';
   }
 
   /// The operator the error happened in
@@ -28,10 +34,11 @@ class MissingOperatorOperandException extends MathException {
 /// Thrown when a function receives less arguments than it expects.
 ///
 /// The function which caused the problem is stored in [func]
-class MissingFunctionArgumentListException extends MathException {
+class MissingFunctionArgumentListException extends MathParseException {
   @override
   String toString() {
-    return 'MissingFunctionArgumentListException: "$func" has insufficient arguments fed';
+    return 'MissingFunctionArgumentListException: "$func" has insufficient '
+        'arguments fed';
   }
 
   /// The function which caused the problem
@@ -46,10 +53,11 @@ class MissingFunctionArgumentListException extends MathException {
 /// Thrown when the parser finds syntax it doesn't understand
 ///
 /// The part which caused the problem is stored in [operation]
-class UnknownOperationException extends MathException {
+class UnknownOperationException extends MathParseException {
   @override
   String toString() {
-    return 'UnknownOperationException: "$operation" is an unknown operation in given context';
+    return 'UnknownOperationException: "$operation" is an unknown operation '
+        'in given context';
   }
 
   /// The part which caused the problem
@@ -64,13 +72,15 @@ class UnknownOperationException extends MathException {
 /// Thrown when some parts of the expression were left unprocessed
 ///
 /// The unprocessed parts are stored as a string in [parts]
-class CantProcessExpressionException extends MathException {
+class CantProcessExpressionException extends MathParseException {
   @override
   String toString() {
     return 'CantProcessExpressionException: '
         'Some parts of the expression were left unprocessed: $parts.'
-        '\nThis often happens if you haven\'t specified the multiplication '
-        'operator explicitly and don\'t have isImplicitMultiplication turned on';
+        '\nThis often happens if you used an undefined variable  in '
+        'variableNames parameter of the parse function or you haven\'t '
+        'specified the multiplication operator explicitly and don\'t have '
+        'isImplicitMultiplication turned on';
   }
 
   /// The unprocessed parts
@@ -85,7 +95,7 @@ class CantProcessExpressionException extends MathException {
 /// Thrown when parsing fails for an unknown reason
 ///
 /// The error is stored in [error]
-class ParsingFailedException extends MathException {
+class ParsingFailedException extends MathParseException {
   @override
   String toString() {
     return 'ParsingFailedException: $error';
@@ -103,10 +113,11 @@ class ParsingFailedException extends MathException {
 /// Thrown when an unexpected closing bracket is met by the parser
 ///
 /// The type of bracket is stored in [type]
-class UnexpectedClosingBracketException extends MathException {
+class UnexpectedClosingBracketException extends MathParseException {
   @override
   String toString() {
-    return 'UnexpectedClosingBracketException: A bracket of type "$type" was found in unexpected context';
+    return 'UnexpectedClosingBracketException: A bracket of type "$type" was '
+        'found in unexpected context';
   }
 
   /// The type of bracket
@@ -122,10 +133,11 @@ class UnexpectedClosingBracketException extends MathException {
 /// correct parsing of the expression
 ///
 /// The type of bracket is stored in [type]
-class BracketsNotClosedException extends MathException {
+class BracketsNotClosedException extends MathParseException {
   @override
   String toString() {
-    return "BracketsNotClosedException: A bracket of type \"$type\" wasn't closed so correct expression parsing can't be guaranteed";
+    return "BracketsNotClosedException: A bracket of type \"$type\" wasn't "
+        "closed so correct expression parsing can't be guaranteed";
   }
 
   /// The type of bracket
@@ -133,4 +145,26 @@ class BracketsNotClosedException extends MathException {
 
   /// Creates a new Brackets Not Closed Exception
   const BracketsNotClosedException(this.type);
+}
+
+/// Invalid Variable Name Exception
+///
+/// Thrown when variable has incorrect name
+///
+/// The name which caused the error is stored in [name]
+class InvalidVariableNameException extends MathParseException {
+  @override
+  String toString() {
+    return "InvalidVariableNameException: A variable with name \"$name\" can't"
+        ' be defined. First character must be a letter, others - letters, '
+        'digits, or underscore. Letters may be latin or Greek, both lower or '
+        'capital case. You can\'t use built-in function names like sin, cos, '
+        'etc. Variable names are case-sensitive';
+  }
+
+  /// The missing variable
+  final String name;
+
+  /// Created a new Undefined Variable Exception
+  const InvalidVariableNameException(this.name);
 }