+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## 1.0.0
+# 1.2.0
+
+- Fix README
+- Moved integrating features to a separate package library `math_parser_integrate`
+
+
+# 1.1.0
+
+- Custom variables support.
+- `MathFunctionX`deprecated.
+- `MathVariable` introduced.
+- You need to pass an instance of `MathVariableValues` instead of a num to the `calc()` function now
+
+
+# 1.0.0
 
 - Initial version.